Bind latestEventID in :in clause when incrementally matching
Previously we were using an equality constraint to ensure the incremental match
queries only considered the latestEventID as a possible event to match. However,
passing this information as part of the :in clause (by modifying the :in clause
to contain the eventLvar of the top-level Winnow clause being executed) should
permit DataScript to make optimizations that improve query performance.