Add preliminary incremental sifting benchmark
This version of the benchmark has a hardcoded partial match pool size
(1000 in this commit), so you need to repeatedly re-run it with a different
value specified in the code to get the full set of results reported in the
Winnow paper draft. I might go back in and change this at some point, so that
you can get a complete set of results from only a single load of the page
and without ever having to change the code.